refactor: moved some functions for integrations into the core package
This will make developing more integrations in future much simpler.
This also involved standardizing the `Options` taken by all integrations

BREAKING CHANGE: `Options` renamed to `ServerOptions` for all integrations
