+++ packages/perseus-actix-web/src/translations.rs
@@ -1,13 +1,14 @@
-use crate::Options;
 use actix_web::{web, HttpRequest, HttpResponse};
 use fmterr::fmt_err;
 use perseus::internal::i18n::TranslationsManager;
+use perseus::internal::serve::ServerOptions;
+use std::rc::Rc;
 
 /// The handler for calls to `.perseus/translations/{locale}`. This will manage returning errors and the like. THe JSON body returned
 /// from this does NOT include the `locale` key, just a `HashMap<String, String>` of the translations themselves.
 pub async fn translations<T: TranslationsManager>(
     req: HttpRequest,
-    opts: web::Data<Options>,
+    opts: web::Data<Rc<ServerOptions>>,
     translations_manager: web::Data<T>,
 ) -> HttpResponse {
     let locale = req.match_info().query("locale");
