+++ packages/perseus/src/html_shell.rs
@@ -1,4 +1,4 @@
-use crate::serve::PageData;
+use crate::page_data::PageData;
 use std::collections::HashMap;
 use std::env;
